fix: update 3D model
